Structure updated
- Navbar changed accordingly
- new Components
- new Project structure
- deleted old Blockchain Components

!!! Functional work still needed: next step !!!
