+++ Webserver/src/components/Core_components/HistoryGraph.js
